% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run ISTbrowser app}
\usage{
run()
}
\description{
Runs \code{shiny::runApp()} on the ISTbrowser shiny app,
as stored in the inst folder within this package.
This function has no arguments.
}
\examples{
\dontrun{run()}

}
